+++ Delta.CertXplorer/src/Delta.CertXplorer/CertManager/Wrappers/X509CertificateWrapper.cs
@@ -16,9 +16,25 @@ namespace Delta.CertXplorer.CertManager.Wrappers
             x509 = certificate;
         }
 
-        public IntPtr Handle { get { return x509.Handle; } }
-        public string Issuer { get { return x509.Issuer; } }
-        public string Subject { get { return x509.Subject; } }
+        public IntPtr Handle { get { return TryGet(() => x509.Handle); } }
+        
+        public string Issuer { get { return TryGet(() => x509.Issuer); } }
+        
+        public string Subject { get { return TryGet(() => x509.Subject); } }
+
+        protected static T TryGet<T>(Func<T> function)
+        {
+            try
+            {
+                return function();
+            }
+            catch (Exception ex)
+            {
+                This.Logger.Error(string.Format("Could not execute function: {0}", ex.Message), ex);
+            }
+
+            return default(T);
+        }
 
     }
 }